Refactor out the 'MenuItem' component
